requirements.txt
fastapi
uvicorn
pymilvus[model]
pydantic-settings
pymongo
bson